refactor: Refactor and Centralize Dependency Resolution Logic
